setting redux to Login, Register
