+++ src/config.js
@@ -1,2 +1,4 @@
 // 공통으로 사용되는 함수
 // ex. API 주소, 이미지 url 등
+
+export const SH = `http://10.58.7.3:8000`;
