+++ src/index.js
@@ -1,16 +1,23 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import Routes from "./Routes";
+import { Provider } from "react-redux";
+import { createStore } from "redux";
 import { ThemeProvider } from "styled-components";
+import rootReducer from "./store/reducers";
 import theme from "./style/theme";
 import GlobalStyle from "./style/global";
 
+const store = createStore(rootReducer);
+
 ReactDOM.render(
   <React.StrictMode>
-    <ThemeProvider theme={theme}>
-      <Routes />
-      <GlobalStyle />
-    </ThemeProvider>
+    <Provider store={store}>
+      <ThemeProvider theme={theme}>
+        <Routes />
+        <GlobalStyle />
+      </ThemeProvider>
+    </Provider>
   </React.StrictMode>,
   document.getElementById("root")
 );
