+++ src/pages/Signup/Signup.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import styled, { css } from "styled-components";
 import Nav from "../../component/Nav/Nav";
