+++ src/store/actions/actionTypes.js
@@ -1,0 +1,19 @@
+// action의 type 정의
+
+export const REGISTER = "REGISTER";
+export const CHECK_EMAIL = "CHECK_EMAIL";
+export const CHECK_PASSWORD = "CHECK_PASSWORD";
+export const CHECK_DOUBLE_PASSWORD = "CHECK_DOUBLE_PASSWORD";
+
+export const LOGIN_EMAIL = "LOGIN_EMAIL";
+export const LOGIN_PASSWORD = "LOGIN_PASSWORD";
+export const LOGOUT = "LOGOUT";
+
+// Exchange
+export const CHANGE_COIN = "CHANGE_COIN";
+export const BUY = "BUY";
+export const SELL = "SELL";
+
+// My Info
+export const CHANGE_MYASSET = "CHANGE_MYASSET";
+export const CHARGE_MONEY = "CHARGE_MONEY";
